Applied header-only approach
