+++ data/User.hpp
@@ -20,17 +20,11 @@ struct User {
      * @param name Nama lengkap user.
      * @param password Password yang digunakan untuk login.
      */
-    User(string id, string name, string password);
-
-    /**
-     * Validasi apakah id dan password yang dimasukkan sesuai dengan akun ini.
-     * 
-     * @param idInput ID yang dimasukkan user.
-     * @param passwordInput Password yang dimasukkan user.
-     * 
-     * @return `true` jika id dan password sesuai dengan akun ini, `false` jika tidak.
-     */
-    bool validate(string idInput, string passwordInput);
+    User(string id, string name, string password) {
+        this->id = id;
+        this->name = name;
+        this->password = password;
+    }
 };
 
 #endif
