--- pages/LoginPage/LoginPage.hpp
@@ -1,8 +1,0 @@
-#ifndef LOGINPAGE_HPP
-#define LOGINPAGE_HPP
-
-
-void handleLogin(string idInput, string passwordInput);
-void LoginPage();
-
-#endif
